requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.15,<3.16
django-cors-headers>=4.3,<5
psycopg2-binary>=2.9
django-filter>=24.2
Pillow>=10.4
python-dotenv>=1.0
xhtml2pdf>=0.2.11
